+++ src/main/java/service/bucketelimination/factor/FactorInstantiate.java
@@ -28,7 +28,7 @@ public class FactorInstantiate {
 
             int oldFactorSize = Util.totalCardinality(f.variables, cardinality);
             int newFactorSize = oldFactorSize / cardinality[variable];
-            nf.factor = new BigDecimal[newFactorSize];
+            nf.factor = new double[newFactorSize];
             f.variables.remove(index);
             nf.variables = f.variables;
             Util.computeStride(nf, cardinality);
